Current state of the code 11-Jan-2017
Basically functionality is there. Plots lineage trees from the .sbd file using a template (currently runs up to 64-cell stage and in code but could be a separate file for flexibilty). Requires the .sbc file in order to determine the image acquisition rate and the left/right delimiter. Needs tidying up and testing with other datasets.